pm_estimates_trigger/requirements.txt
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
google-api-core==1.31.0
google-auth==1.32.1
google-cloud-core==1.7.1
google-cloud-pubsub==2.6.1
google-cloud-storage==1.40.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.38.1
idna==2.10
libcst==0.3.19
mypy-extensions==0.4.3
packaging==21.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.6
